[core] Add compression for ExternalBuffer
